Updated: Antihack (REWORK REQUIRED)
Requires rework to see which section is causing setdamage 1 not to
properly spawn player death.

In this patch I've fixed teleport reverting from debug, added some heavy
diag logging incase this is causing issues still though, no toggle for
it yet.

Also made it so the banned commands should tell the server to kick the
player off and log them to the hacker log.

Added a toggle to the init sqf that gets read in here to turn it on/off
for admins.

For debug distance checking I've used killzone kids function as it's
more accurate over large distances.