Static Vehicle Spawn Mechanic Update
Now counts existing non-char-bought vehicles, and if less than the (new)
amount to spawn count, it will spawn up to the max count at randomised
position selections from the list. The position randomisation is also
new, the old mechanic spawned it at first position, then second, etc.
If it finds a vehicle already at the spawn point it will fail, it can
fail a max of 25 times per vehicle before it gives up trying to place it
and continues to the next vehicle. This mechanic addition is also new.
Another new addition to an existing mechanic is that 350'z colour now
gets randomised as well as suv colour.